CLOUD_RUN/requirements.txt
pymongo==3.13.0
pymongo[srv]
python-dotenv==1.0.1
fastapi==0.114.0
uvicorn==0.30.6
google-cloud-pubsub==2.23.0
gunicorn==22.0.0